feat: handle authorization and complete profile router

+++ routes/orders.router.js
@@ -1,31 +1,37 @@
 const { Router } = require('express');
+const passport = require('passport')
 
 const {
-  createOrderSchema,
-  updateOrderSchema,
   getOrderSchema,
   addItemSchema,
   updateItemSchema,
   getItemSchema,
 } = require('../schemas/order.schema');
 const validatorHandler = require('../middlewares/validator.handler');
+const { checkOwner, checkRoles } = require('../middlewares/auth.handler')
 const OrderService = require('../services/order.service');
 
 const router = Router();
 const service = new OrderService();
 
-router.get('/', async (req, res, next) => {
-  try {
-    const orders = await service.find();
-    res.json(orders);
-  } catch (error) {
-    next(error);
-  }
+router.get(
+  '/',
+  passport.authenticate('jwt', { session: false }),
+  checkRoles('admin'),
+  async (req, res, next) => {
+    try {
+      const orders = await service.find();
+      res.json(orders);
+    } catch (error) {
+      next(error);
+    }
 });
 
 router.get(
   '/:id',
+  passport.authenticate('jwt', { session: false }),
   validatorHandler(getOrderSchema, 'params'),
+  checkOwner('Order'),
   async (req, res, next) => {
     try {
       const { id } = req.params;
@@ -38,23 +44,11 @@ router.get(
 );
 
 router.post(
-  '/',
-  validatorHandler(createOrderSchema, 'body'),
-  async (req, res, next) => {
-    try {
-      const data = req.body;
-      const newOrder = await service.create(data);
-      res.status(201).json(newOrder);
-    } catch (error) {
-      next(error);
-    }
-  },
-);
-
-router.post(
   '/:id/products',
+  passport.authenticate('jwt', { session: false }),
   validatorHandler(getOrderSchema, 'params'),
   validatorHandler(addItemSchema, 'body'),
+  checkOwner('Order'),
   async (req, res, next) => {
     try {
       const { id } = req.params;
@@ -68,28 +62,14 @@ router.post(
 );
 
 router.patch(
-  '/:id',
-  validatorHandler(getOrderSchema, 'params'),
-  validatorHandler(updateOrderSchema, 'body'),
-  async (req, res, next) => {
-    try {
-      const { id } = req.params;
-      const body = req.body;
-      const updatedOrder = await service.update(id, body);
-      res.json(updatedOrder);
-    } catch (error) {
-      next(error);
-    }
-  },
-);
-
-router.patch(
-  '/:orderId/products/:productId',
+  '/:id/products/:productId',
+  passport.authenticate('jwt', { session: false }),
   validatorHandler(getItemSchema, 'params'),
   validatorHandler(updateItemSchema, 'body'),
+  checkOwner('Order'),
   async (req, res, next) => {
     try {
-      const { orderId, productId } = req.params;
+      const { id: orderId, productId } = req.params;
       const data = req.body;
       const newItem = await service.updateItem(orderId, productId, data);
       res.json(newItem);
@@ -101,7 +81,9 @@ router.patch(
 
 router.delete(
   '/:id',
+  passport.authenticate('jwt', { session: false }),
   validatorHandler(getOrderSchema, 'params'),
+  checkOwner('Order'),
   async (req, res, next) => {
     try {
       const { id } = req.params;
@@ -114,11 +96,13 @@ router.delete(
 );
 
 router.delete(
-  '/:orderId/products/:productId',
+  '/:id/products/:productId',
+  passport.authenticate('jwt', { session: false }),
   validatorHandler(getItemSchema, 'params'),
+  checkOwner('Order'),
   async (req, res, next) => {
     try {
-      const { orderId, productId } = req.params;
+      const { id: orderId, productId } = req.params;
       const deletedItem = await service.deleteItem(orderId, productId);
       res.json(deletedItem);
     } catch (error) {
